requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
influxdb==5.2.0
kafka-python==1.4.2
python-dateutil==2.7.3
pytz==2018.5
requests>=2.20.0
urllib3>=1.24.2
six==1.11.0